Control.Functor.HT

Documentation

void :: Functor f => f a -> f ()

map :: Functor f => (a -> b) -> f a -> f b

for :: Functor f => f a -> (a -> b) -> f b

unzip :: Functor f => f (a, b) -> (f a, f b)

Caution: Every pair member has a reference to the argument of unzip. Depending on the consumption pattern this may cause a memory leak. For lists, I think, you should generally prefer unzip.

unzip3 :: Functor f => f (a, b, c) -> (f a, f b, f c)

Caution: See unzip.

uncurry :: Functor f => (f a -> f b -> g) -> f (a, b) -> g

Caution: See unzip.

uncurry3 :: Functor f => (f a -> f b -> f c -> g) -> f (a, b, c) -> g

Caution: See unzip.

mapFst :: Functor f => (a -> f c) -> (a, b) -> f (c, b)

mapSnd :: Functor f => (b -> f c) -> (a, b) -> f (a, c)

mapFst3 :: Functor f => (a -> f d) -> (a, b, c) -> f (d, b, c)

mapSnd3 :: Functor f => (b -> f d) -> (a, b, c) -> f (a, d, c)

mapThd3 :: Functor f => (c -> f d) -> (a, b, c) -> f (a, b, d)

outerProduct :: (Functor f, Functor g) => (a -> b -> c) -> f a -> g b -> f (g c)

Generalization of outerProduct.