Data.Ord.HT

Documentation

comparing :: Ord b => (a -> b) -> a -> a -> Ordering

limit :: Ord a => (a, a) -> a -> a

limit (lower,upper) x restricts x to the range from lower to upper. Don't expect a sensible result for lower>upper.

inRange :: Ord a => (a, a) -> a -> Bool

limit (lower,upper) x checks whether x is in the range from lower to upper. Don't expect a sensible result for lower>upper.