Data.Record.HT

Documentation

compare :: [a -> a -> Ordering] -> a -> a -> Ordering

Lexicographically compare a list of attributes of two records.

Example:

compare [comparing fst, comparing snd]

equal :: [a -> a -> Bool] -> a -> a -> Bool

Check whether a selected set of fields of two records is equal.

Example:

equal [equating fst, equating snd]